fix added2repo timestamp by metadata (if we have them)
The F-Droid repo only keeps the "added" date at the date of the latest
APK file existing in the repo - so once that gets purged, it's no
longer correct. If we have the metadata available, we can fix this up
at least by the file date (if older), or using the "AppAdded:" line
if found in the "MaintainerNotes" section.